Work correctly when enhanced nav is disabled
